tests: fix file clear with procps 3.16 (LP: #1864423)
With the new procps installed the current construct to clear files in
the tests fails to clear the log and due to that breaks later checks.
Swicth to the more reliable (and more readable) truncate to resolve the
issue.

More details at https://bugs.launchpad.net/ubuntu/+source/procps/+bug/1864423

Signed-off-by: Christian Ehrhardt <christian.ehrhardt@canonical.com>