dh_make_pgxs: Use current debhelper version instead of the oldest one.
We used to keep the debhelper compat level to the lowest supported in all
Debian and Ubuntu releases, but the span is getting too large so scripts have
been put in place to auto-set the compat level for older releases in the
apt.postgresql.org buildds. This works as long as no new features (like
execute_after_*) are being used.