server/postgresql.mk: avoid gcc 11 ICE on armhf and armel
Currently the builds of 14.1-3 are failing [1][2]:

during RTL pass: mach
/<<PKGBUILDDIR>>/build/../src/backend/storage/buffer/bufmgr.c: In function ‘ReadBuffer_common’:
/<<PKGBUILDDIR>>/build/../src/backend/storage/buffer/bufmgr.c:1085:1: internal compiler error: in create_fix_barrier, at config/arm/arm.c:17891
 1085 | }
      | ^
0x106b1cf internal_error(char const*, ...)
	???:0
0x19f163 fancy_abort(char const*, int, char const*)
	???:0

This is due to a bug in gcc-11.2 [3] and will eventually be resolved
there. But for now to get builds going again we have to work around it.

This is the submission of a similar (before centralized d/rules) fix
that was uploaded to Ubuntu [4] postgresql-14.

[1]: https://buildd.debian.org/status/fetch.php?pkg=postgresql-14&arch=armhf&ver=14.1-3&stamp=1638530270&raw=0
[2]: https://buildd.debian.org/status/fetch.php?pkg=postgresql-14&arch=armel&ver=14.1-3&stamp=1638530561&raw=0
[3]: https://gcc.gnu.org/bugzilla/show_bug.cgi?id=103395
[4]: https://launchpad.net/ubuntu/+source/postgresql-14/1